+++ src/viewModels/TextModel.ts
